Improve serialization, change rate logic
After a peer programming session with @gre we found a way to serialize/deserialize even for token/subaccounts. Still pending the integration on desktop.
This also includes the rate logic where we divide by the 10^fromUnit.magnitude in order to make the rates be between satoshi units instead of top units.