Merge Fees in other Operations
The Fees aren't included in the operations list given by Ledger Live
Common. A fold is accomplished over the list of transactions given by
LibCore that includes the fees in all the other operations that match
the transaction Hash, and therefore the fees operations can be ignored
in the end result.