Add getAppAndVersion step to connectmanager
In order to have a more reliable connectManager command response handling we need this extra step since the device response to the getVersion apdu is not consistent. In particular, it collided with the Exchange app which considers the second byte 01 to be a swap process start and was wrongly interpreted as device in bootloader case. This addresses that.