Adapt Swap History to erc20 tokens without funds yet
Allow swap operations for tokens that dont have funds yet to still show on the list as if we had an account for them. This is done by enhancing the main account and checking if the token account already existed in the subaccount list. We need this in order to prevent navigation to an account that doesn't really exist in our data.