Fixes libcore to really support changing config
there is a libcore issue that makes this scenario not working:
getWallet --> updateWalletConfig --> getWallet

instead what this is doing is:
try { return createWallet } catch { updateWallet };  return getWallet

Scenario i was using is:

	VERBOSE=1 ledger-live sync --id dash:xpubhere
	EXPERIMENTAL_EXPLORERS=1 ledger-live sync --id dash:xpubhere
	VERBOSE=1 ledger-live sync --id dash:xpubhere

at any point in time it needs to use v2 or v3 but not stay on the first config used. this PR fixes it.